Refactor shared TIMES timeslice helpers
Pull the duplicated timeslice construction logic out of PREPARE-TIMES-NZ/scripts/stage_1_prep_raw_data/extract_ea_data.py and PREPARE-TIMES-NZ/scripts/stage_3_scenarios/electricity/
solar_run_hourly_profiles.py into a new shared module, PREPARE-TIMES-NZ/src/prepare_times_nz/utilities/timeslices.py.

Point PREPARE-TIMES-NZ/src/prepare_times_nz/stage_2/load_curves/ residential_curves.py at the same shared hour-to-time-of-day helper, since it had another copy of that mapping logic.

Behavior should stay the same. Change was to remove duplicate code that pylint was failing on and keep the mapping defined in one place.
